+++ src/components/Overlay.vue
@@ -8,12 +8,10 @@ const { props = {} } = defineProps<{
 
 <template>
   <div :class="C.Overlay.overlay">
-    <div :class="[C.Overlay.background, C.Overlay.overlay]">
-      <div :class="C.Overlay.close" @click="onClose" />
-      <div :class="C.Overlay.children">
-        <Component :is="child" v-bind="props" @close="onClose" />
-      </div>
-      <div :class="C.Overlay.close" @click="onClose" />
+    <div :class="C.Overlay.close" @click="onClose" />
+    <div :class="C.Overlay.children">
+      <Component :is="child" v-bind="props" @close="onClose" />
     </div>
+    <div :class="C.Overlay.close" @click="onClose" />
   </div>
 </template>