+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Fixed
 
 - `mtra-transfer-on-enter`: Fix feature not working for docked tiles
+- Fix overlays not showing up
 
 ## 25.2.6.1805