+++ CHANGELOG.md
@@ -1,9 +1,12 @@
 ## Unreleased
 
+### Added
+
+- `prod-hide-percent`: New feature. Hides percent value from production lines.
+
 ### Fixed
 
 - Fix duplication of Materials in Transit asset value in Long-Term Materials Receivable
-- `prod-hide-percent`: New feature. Hides percent value from production lines.
 
 ## 25.2.11