+++ CHANGELOG.md
@@ -2,8 +2,10 @@
 
 ### Added
 
-- `context-controls-no-hover`: New feature. Prevents the context controls from displaying description while hovering over
+- `XIT SHEET`: Add an optional parameter for Sheet ID
 - `cxos-leftalign-delete-filled`: New feature. Moves 'Delete Filled' button to the left in CXOS
+- `context-controls-no-hover`: New feature. Prevents the context controls from displaying description while hovering
+  over
 - `prod-hide-percent`: New feature. Hides percent value from production lines.
 
 ### Fixed