+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - `XIT SHEET`: Add an optional parameter for Sheet ID
 - `context-controls-no-hover`: New feature. Prevents the context controls from displaying description while hovering
   over
+- `inv-compress-inventory-info`: New feature. Compresses specific inventory info into a row
 - `prod-hide-percent`: New feature. Hides percent value from production lines.
 
 ### Changed