+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## Unreleased
+## 25.2.25
 
 ### Added